% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourPNO-package.R
\docType{package}
\name{fourPNO-package}
\alias{fourPNO}
\alias{fourPNO-package}
\title{fourPNO: Bayesian 4 Parameter Item Response Model}
\description{
Estimate Barton & Lord's (1981) <doi:10.1002/j.2333-8504.1981.tb01255.x> 
    four parameter IRT model with lower and upper asymptotes using Bayesian
    formulation described by Culpepper (2016) <doi:10.1007/s11336-015-9477-6>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tmsalab/fourPNO}
  \item Report bugs at \url{https://github.com/tmsalab/fourPNO/issues}
}

}
\author{
\strong{Maintainer}: Steven Andrew Culpepper \email{sculpepp@illinois.edu} (0000-0003-4226-6176) [copyright holder]

}
\keyword{internal}
