% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kappa_initialize}
\alias{kappa_initialize}
\title{Initialize Thresholds}
\usage{
kappa_initialize(Ms)
}
\arguments{
\item{Ms}{A \code{vector} with the number of scale values.}
}
\value{
A \code{matrix} that is a Multivariate Normal distribution
}
\description{
Internal function for initializing item thresholds.
}
\author{
Steven Andrew Culpepper
}
\seealso{
\code{\link{Gibbs_4PNO}}
}

