% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{min2LL_4pno}
\alias{min2LL_4pno}
\title{Compute 4PNO Deviance}
\usage{
min2LL_4pno(N, J, Y, as, bs, gs, ss, theta)
}
\arguments{
\item{N}{An \code{int}, which gives the number of observations.  (> 0)}

\item{J}{An \code{int}, which gives the number of items.  (> 0)}

\item{Y}{A N by J \code{matrix} of item responses.}

\item{as}{A \code{vector} of item discrimination parameters.}

\item{bs}{A \code{vector} of item threshold parameters.}

\item{gs}{A \code{vector} of item lower asymptote parameters.}

\item{ss}{A \code{vector} of item upper asymptote parameters.}

\item{theta}{A \code{vector} of prior thetas.}
}
\value{
-2LL.
}
\description{
Internal function to -2LL
}
\seealso{
\code{\link[=Gibbs_4PNO]{Gibbs_4PNO()}}
}
\author{
Steven Andrew Culpepper
}
