% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Total_Tabulate}
\alias{Total_Tabulate}
\title{Calculate Tabulated Total Scores}
\usage{
Total_Tabulate(N, J, Y)
}
\arguments{
\item{N}{An \code{int}, which gives the number of observations. (> 0)}

\item{J}{An \code{int}, which gives the number of items. (> 0)}

\item{Y}{A N by J \code{matrix} of item responses.}
}
\value{
A vector of tabulated total scores.
}
\description{
Internal function to -2LL
}
\seealso{
\code{\link[=Gibbs_4PNO]{Gibbs_4PNO()}}
}
\author{
Steven Andrew Culpepper
}
