% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_2pno}
\alias{update_2pno}
\title{Update 2PNO Model Parameters}
\usage{
  update_2pno(N,J,Y,Z,as,bs,theta,Kaps,mu_xi,Sigma_xi_inv,mu_theta,Sigma_theta_inv)
}
\arguments{
\item{N}{The number of observations.}

\item{J}{The number of items.}

\item{Y}{A N by J \code{matrix} of item responses.}

\item{Z}{A \code{matrix} N by J of continuous augmented data.}

\item{as}{A \code{vector} of item discrimination parameters.}

\item{bs}{A \code{vector} of item threshold parameters.}

\item{theta}{A \code{vector} of prior thetas.}

\item{Kaps}{A \code{matrix} for item thresholds (used for internal computations).}

\item{mu_xi}{Prior mean for item parameters.}
\item{Sigma_xi_inv}{Prior item parameter inverse variance-covariance matrix.}
\item{mu_theta}{Prior mean for theta.}
\item{Sigma_theta_inv}{Prior inverse variance for theta.}
}
\value{
A list of 2PNO model parameters.
}
\description{
Internal function to update 2PNO model parameters
}
\author{
Steven Andrew Culpepper
}
\seealso{
\code{\link{Gibbs_4PNO}}
}

