% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_ab_norestriction}
\alias{update_ab_norestriction}
\title{Update a and b Parameters of 4pno without alpha > 0 Restriction}
\usage{
update_ab_norestriction(N, J, Z, as, bs, theta, mu_xi, Sigma_xi_inv)
}
\arguments{
\item{N}{An \code{int}, which gives the number of observations.  (> 0)}

\item{J}{An \code{int}, which gives the number of items.  (> 0)}

\item{Z}{A \code{matrix} N by J of continuous augmented data.}

\item{as}{A \code{vector} of item discrimination parameters.}

\item{bs}{A \code{vector} of item threshold parameters.}

\item{theta}{A \code{vector} of prior thetas.}

\item{mu_xi}{A two dimensional \code{vector} of prior item parameter means.}

\item{Sigma_xi_inv}{A two dimensional identity \code{matrix} of prior item parameter VC matrix.}
}
\value{
A list of item parameters.
}
\description{
Update item slope and threshold
}
\author{
Steven Andrew Culpepper
}
\seealso{
\code{\link{Gibbs_4PNO}}
}

