% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_theta}
\alias{update_theta}
\title{Internal Function for Updating Theta in Gibbs Sampler}
\usage{
update_theta(N, Z, as, bs, theta, mu_theta, Sigma_theta_inv)
}
\arguments{
\item{N}{An \code{int}, which gives the number of observations.  (> 0)}

\item{Z}{A \code{matrix} N by J of continuous augmented data.}

\item{as}{A \code{vector} of item discrimination parameters.}

\item{bs}{A \code{vector} of item threshold parameters.}

\item{theta}{A \code{vector} of prior thetas.}

\item{mu_theta}{The prior mean for theta.}

\item{Sigma_theta_inv}{The prior inverse variance for theta.}
}
\value{
A \code{vector} of thetas.
}
\description{
Update theta in Gibbs sampler
}
\author{
Steven Andrew Culpepper
}
\seealso{
\code{\link{Gibbs_4PNO}}
}

