% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_WKappaZ_NA}
\alias{update_WKappaZ_NA}
\title{Update Lower and Upper Asymptote Parameters of 4PNO}
\usage{
update_WKappaZ_NA(Y, Ysum, Z, as, bs, gs, ss, theta, Kaps, alpha_c, beta_c,
  alpha_s, beta_s, gwg_reps)
}
\arguments{
\item{Y}{A N by J \code{matrix} of item responses.}

\item{Ysum}{A \code{vector} of item total scores.}

\item{Z}{A \code{matrix} N by J of continuous augmented data.}

\item{as}{A \code{vector} of item discrimination parameters.}

\item{bs}{A \code{vector} of item threshold parameters.}

\item{gs}{A \code{vector} of item lower asymptote parameters.}

\item{ss}{A \code{vector} of item upper asymptote parameters.}

\item{theta}{A \code{vector} of prior thetas.}

\item{Kaps}{A \code{matrix} for item thresholds (used for internal computations).}

\item{alpha_c}{The lower asymptote prior 'a' parameter.}

\item{beta_c}{The lower asymptote prior 'b' parameter.}

\item{alpha_s}{The upper asymptote prior 'a' parameter.}

\item{beta_s}{The upper asymptote prior 'b' parameter.}

\item{gwg_reps}{The number of Gibbs within Gibbs MCMC samples for marginal distribution of gamma.}
}
\value{
A list of item threshold parameters.
}
\description{
Internal function to update item lower and upper asymptote
}
\author{
Steven Andrew Culpepper
}
\seealso{
\code{\link{Gibbs_4PNO}}
}

