% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Y_4pno_simulate}
\alias{Y_4pno_simulate}
\title{Simulate from 4PNO Model}
\usage{
Y_4pno_simulate(N, J, as, bs, gs, ss, theta)
}
\arguments{
\item{N}{An \code{int}, which gives the number of observations. (> 0)}

\item{J}{An \code{int}, which gives the number of items. (> 0)}

\item{as}{A \code{vector} of item discrimination parameters.}

\item{bs}{A \code{vector} of item threshold parameters.}

\item{gs}{A \code{vector} of item lower asymptote parameters.}

\item{ss}{A \code{vector} of item upper asymptote parameters.}

\item{theta}{A \code{vector} of prior thetas.}
}
\value{
A N by J \code{matrix} of dichotomous item responses.
}
\description{
Generate item responses under the 4PNO
}
\seealso{
\code{\link[=Gibbs_4PNO]{Gibbs_4PNO()}}
}
\author{
Steven Andrew Culpepper
}
